requirements.txt
terra_sdk
requests
python-dotenv